Refactor bundler into interface
This is to allow easier implementation of alternate bundling strategies
