+++ src/main/java/com/shapesecurity/bandolier/Main.java
@@ -1,5 +1,6 @@
 package com.shapesecurity.bandolier;
 
+import com.shapesecurity.bandolier.bundlers.StandardModuleBundler;
 import com.shapesecurity.bandolier.loader.FileLoader;
 import com.shapesecurity.bandolier.loader.IResourceLoader;
 import com.shapesecurity.bandolier.loader.NodeResolver;
@@ -14,7 +15,7 @@ public class Main {
 			IResourceLoader loader = new FileLoader();
 			Script bundle = Bundler.bundle(Paths.get(args[0]).toAbsolutePath(),
 										   new NodeResolver(loader),
-										   loader);
+										   loader, new StandardModuleBundler());
 			System.out.print(CodeGen.codeGen(bundle));
 		} else {
 			System.err.println("Must provide a filename");
